backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.3.1

pandas==2.1.4
numpy==1.26.4

openpyxl==3.1.2
reportlab==4.0.8

python-decouple==3.8
requests==2.31.0
